Implement DVCBS

+++ pancake/GBFHS.h
@@ -138,7 +138,7 @@ class Gbfhs {
     }
 
     expansions = 0;
-    UB = std::numeric_limits<double>::max();
+    UB = std::numeric_limits<size_t>::max();
 
     size_t iteration = 0;
